Improve display of messages.
After a lot of fscking around I've finally gotten the messages look
(almost) the way I want to.  Adding a separator between them does not
work yet, but each message is on one line now with the right margins.

QML feels a lot like CSS with turning and tweaking knobs in different
places.