Get rid of `ecl-quicklisp` when compiling app.
This was the main culprit of the slow loading time and of the initial
huge loading time since it installed Quicklisp and all the required
packages AGAIN.

Now it starts up quicker.  Not quick enough to my liking, but
acceptable enough to drop the priority of this issue down my to-do
list.