Disable notifications.
At the moment notifications are not useful, just annoying, since the
app needs to be open to retrieve messages anyway *and* new messages
will be shown on the cover.

Once the app can be closed and messages retrieved in the background
this can be enabled again (but with a toggle in the config!).